requirements.txt
#pip install -r requirements.txt
Pillow
requests
numpy
pytz
piexif
python-dotenv
opencv-python